Add BriX for building data structures
We developed a small library which you can use to build data objects.
These objects can be printed to media, following the printer pattern. The objects available in the library are a format-feature-merge of XML and JSON so when you use the library, you ensure that these media types are supported.